Add problems A, D, and E
Add problems A, D, E and updated files.

+++ D-CropCirclesAndTryangles/logic.c
@@ -6,8 +6,8 @@ int main()
 	scanf("%d",&T);
 	while(T--)
 	{
-		long long r;
-		scanf("%lld",&r);
+		long r;
+		scanf("%ld",&r);
 		printf("%0.2lf\n",r*r*((((4*1.73 - 5)*3.14) - 3*1.73)/2));		//(((4*1.73 - 5)*3.14) - 3*1.73)/2 = 0.4194
 	}
 	return 0;
